Rename armtalker_api to arm_client
