+++ src/arm_client/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/arm_client
+[install]
+install-scripts=$base/lib/arm_client
