--- src/armtalker_api/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/armtalker_api
-[install]
-install-scripts=$base/lib/armtalker_api
