+++ src/rostalker2/package.xml
@@ -13,7 +13,7 @@
   <test_depend>python3-pytest</test_depend>
 
   <exec_depend>rostalker2interface</exec_depend>
-  <exec_depend>armtalker_api</exec_depend>
+  <exec_depend>arm_client</exec_depend>
   <exec_depend>mastertalker_api</exec_depend>
 
   <export>
